requirements.txt
-i https://pypi.org/simple
anyio==3.7.0 ; python_version >= '3.7'
black==23.3.0
click==8.1.3 ; python_version >= '3.7'
fastapi==0.97.0
h11==0.14.0 ; python_version >= '3.7'
httptools==0.5.0
idna==3.4 ; python_version >= '3.5'
mypy-extensions==1.0.0 ; python_version >= '3.5'
packaging==23.1 ; python_version >= '3.7'
pathspec==0.11.1 ; python_version >= '3.7'
platformdirs==3.5.3 ; python_version >= '3.7'
pydantic==1.10.9 ; python_version >= '3.7'
python-dotenv==1.0.0
pyyaml==6.0
sniffio==1.3.0 ; python_version >= '3.7'
starlette==0.27.0 ; python_version >= '3.7'
typing-extensions==4.6.3 ; python_version >= '3.7'
uvicorn[standard]==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.3
